compatible/v0/requirements.txt
adal==1.1.0
ansible==2.6.0
Babel==0.9.6
backoff==1.0.7
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
cachetools==2.1.0
certifi==2018.4.16
cffi==1.6.0
chardet==3.0.4
click==6.6
configobj==4.7.2
cryptography==1.7.2
decorator==3.4.0
dictdiffer==0.7.1
docker-py==1.10.6
docker-pycreds==0.3.0
docopt==0.6.2
enum34==1.0.4
google-auth==1.5.1
hiredis==0.2.0
httplib2==0.9.2
idna==2.7
iniparse==0.4
ipaddr==2.1.11
ipaddress==1.0.22
IPy==0.75
ipython==5.8.0
ipython-genutils==0.2.0
javapackages==1.0.0
Jinja2==2.7.2
jmespath==0.9.0
json2yaml==1.1.1
kitchen==1.1.1
kong==0.2.1
kubernetes==7.0.0
libvirt-python==3.9.0
lxml==3.2.1
M2Crypto==0.21.1
MarkupSafe==0.11
oauthlib==2.1.0
openshift==0.7.1
ordereddict==1.1
paramiko==2.1.1
passlib==1.6.5
pathlib2==2.3.2
perf==0.1
pexpect==4.6.0
pickleshare==0.7.4
ply==3.4
policycoreutils-default-encoding==0.1
poster==0.8.1
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pyaml==17.12.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.14
pycrypto==2.6.1
pycurl==7.19.0
Pygments==2.2.0
pygobject==3.22.0
pygpgme==0.3
PyJWT==1.6.4
pyliblzma==0.5.3
python-dateutil==2.7.3
python-keyczar==0.71rc0
python-kong==0.4.0
python-linux-procfs==0.4.9
python-string-utils==0.6.0
pyudev==0.15
pyxattr==0.5.1
PyYAML==3.13
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
rsa==3.4.2
ruamel.ordereddict==0.4.13
ruamel.yaml==0.15.66
scandir==1.9.0
schedutils==0.4
seobject==0.1
sepolicy==1.1
simplegeneric==0.8.1
six==1.11.0
slip==0.4.0
slip.dbus==0.4.0
suds==0.4.1
templated-yaml==1.1.1
traitlets==4.3.2
urlgrabber==3.10
urllib3==1.23
virt-who==0.21.7
virtualenv==16.0.0
wcwidth==0.1.7
websocket-client==0.48.0
yum-metadata-parser==1.1.4
